connect QuanaDeviceCommunicatorCallback, add timer for scan and upload, add recyclerView for sampleResponseData

+++ app/src/main/java/il/co/quana/model/GeneralModel.kt
@@ -37,7 +37,7 @@ data class SampleResponse(
 
 data class SampleResponseData(
     val name: String?,
-    val Percentage: Int?
+    val Percentage: Double?
 )
 
 enum class SampleStatus{
